requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.28
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
